pyproject.toml
[project]
name = "documind"
version = "0.1.0"
description = "DocuMind — A production-grade RAG chatbot for your documents."
authors = [{ name = "Jayasoruban R" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "streamlit>=1.40.0",
    "langchain>=0.3.14",
    "langchain-core>=0.3.29",
    "langchain-community>=0.3.14",
    "langchain-openai>=0.3.0",
    "langchain-chroma>=0.2.0",
    "langchain-text-splitters>=0.3.5",
    "chromadb>=0.5.23",
    "openai>=1.60.0",
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.8.0",
    "unstructured>=0.16.11",
    "markdown>=3.7",
    "opentelemetry-exporter-otlp-proto-grpc>=1.30.0",
    "opentelemetry-proto>=1.30.0",
    "protobuf>=5.0.0,<6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
